pyproject.toml
[project]
name = "marksclock"
version = "0.1.0"
description = "Exhaustive time utility suite — clock, timers, world clock, converters, and more"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "astral>=3.2",
    "python-dateutil>=2.9.0",
    "click>=8.1.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
marksclock = "marksclock.__main__:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
